Fix subscription timing and D1 bigint error
Pylon CLI:
- Move job/chat subscriptions to after auth (relay requires it)
- Fix cli.rs to remove debug logging

Relay Worker:
- Fix D1 bigint error by using f64 instead of i64 for kind/created_at
  (JavaScript BigInt not supported by D1)

Test:
- Add send_job_test.rs for E2E job testing

E2E test now passes:
- CLI connects to FM Bridge and authenticates with relay
- CLI receives job requests from network
- CLI processes jobs through FM inference
- CLI publishes results and increments credits

+++ crates/relay-worker/src/storage.rs
@@ -55,8 +55,9 @@ impl<'a> Storage<'a> {
             .bind(&[
                 event.id.clone().into(),
                 event.pubkey.clone().into(),
-                (event.kind as i64).into(),
-                (event.created_at as i64).into(),
+                // Use f64 instead of i64 for D1 compatibility (BigInt not supported)
+                (event.kind as f64).into(),
+                (event.created_at as f64).into(),
                 event.content.clone().into(),
                 tags_json.into(),
                 event.sig.clone().into(),
